python-download-all-invoices/requirements.txt
requests==2.32.3
python_dotenv==1.0.1